refactor: Controller 패키지 삭제, 클래스 이름 변경 (controller->GamePlay)
