+++ src/main/java/blackjack/Application.java
@@ -1,11 +1,9 @@
 package blackjack;
 
-import blackjack.controller.BlackjackController;
-
 public class Application {
 
     public static void main(String[] args) {
-        BlackjackController blackjackController = new BlackjackController();
-        blackjackController.run();
+        BlackjackGamePlay blackjackGamePlay = new BlackjackGamePlay();
+        blackjackGamePlay.run();
     }
 }
